Pull up and reuse ClusterId
